将 用户配置文件 由 本地配置目录 改为 任意目录，PackageManager中添加了从用户配置文件获取Bundles的UI部分。
- 为所有模块的 新建配置前 添加 用户配置根目录 可用性检查
- 移除了本地目录下的所有用户配置

+++ Assets/_BuildPipeline/EazyBuildPipeline/AssetPreprocessor/Editor/SettingPanel.cs
@@ -238,10 +238,15 @@ namespace EazyBuildPipeline.AssetPreprocessor.Editor
 
 		private void CreateNewUserConfig(string name, string path)
 		{
-			//新建   
-			Directory.CreateDirectory(Path.GetDirectoryName(path));
+            //新建
+            if (!Directory.Exists(CommonModule.CommonConfig.Json.UserConfigsRootPath))
+            {
+                G.Module.DisplayDialog("创建失败！用户配置根目录不存在：" + CommonModule.CommonConfig.Json.UserConfigsRootPath);
+                return;
+            }
+            Directory.CreateDirectory(Path.GetDirectoryName(path));
 			File.Create(path).Close();
-			EditorUtility.DisplayDialog("创建成功", "创建成功!", "确定");
+			G.Module.DisplayDialog("创建成功!");
 			//更新列表
 			userConfigNames = EBPUtility.FindFilesRelativePathWithoutExtension(G.Module.ModuleConfig.UserConfigsFolderPath);
 			//保存
